....I..... [ZBX-23441] fixed go install auto{conf,make} integration
This gets rid of setting the $GOBIN environment globally and sets it at
the time of invocation for `go install`. Setting $GOBIN in the autoconf
template overlooked the fact that $DESTDIR may be set, and would thus
always install the applications to the system application directory
regardless. This is done on a per-application basis, as not all
applications should end up in the system application directory.