....I..... [ZBX-23441] removed setting of GOCMD and make use of GO
This removes the explicit setting of GOCMD, which was never used, and
replaces the substitution of it for substituting GO in makefiles, as
well as replacing bare invocations of `go` with whatever $GO now holds.

The rename is to be consistent with autotool conventions of calling the
environment variable the same as the expected utility that performs the
function. This is also what autoconf sets (as opposed to GOCMD) in its
generic program tests.