.......... [ZBX-23441] updated install targets for zabbix_{agent2,web_service}
Previously, we relied on $(GO) to install, and if necessary to build,
the executables when making the install targets. This was to simply
drive the Go tooling via autotools.

As this was a little a tad verbose, and the install targets (in the
agent case) also did additional things (such as installing configuration
files), simply rely on the build target instead and install it.

The configuration installation has also changed. Previously, we tested
whether a configuration file at the destination already existed, and
only if it did not, would we install the file. This is the opposite of
what autotools do, and it is the users responsibility to ensure
configuration backups.

Additionally, the variable AGENT2_CONFIG_PLUGINS_PATH has been renamed
to AGENT2_PLUGIN_CONFIG_DIR, as a _PATH suffix doesn't distinguish
betwween a file and a directory, while we're generally using _FILE to
indicate a file.