...G...... [ZBX-24389] fixed make not rebuilding go executables
This was caused by ZBX-23441, which changed the target names for
building the executables. Previously, the targets would build
executables and output them to a different location than the actual
target name, thus they would be rebuilt on each invocation of make.

Once the target names were changed to the actual names as they appear on
the filesystem, make would no longer rebuild these targets after doing
so once, as they had no prerequisites and if the file existed, the
target was considered up-to-date.